requirements.txt
langchain
langchain-community
langchain-openai
openai
faiss-cpu
PyMuPDF
python-docx
streamlit
aiohttp
nest_asyncio
tavily-python
requests
python-dotenv
pandas
tqdm
pydantic